Split of Object file to multiple sources
Split Object to ObjectBase, Object, ObjectRectangle. Curruntly AM uses ObjectRectangle
for storing objects. In case of AI it may require pixels storage as Object.cpp has, swithcing back is possible.

Cmake: Add compiling of static library for aqumarine. Then this library linking to aquamarine demo.
